[coverage] added machinery for local coverage computation

+++ build-docs.sh
@@ -1,15 +1,21 @@
 #!/bin/bash
 
 die () { echo "ERROR: $*" >&2; exit 2; }
+warn() { echo "WARNING: $*"; }
 
 for cmd in mkdocs pdoc3; do
     command -v "$cmd" >/dev/null ||
-        die "Missing $cmd; \`pip install $cmd\`"
+        warn "Missing $cmd; \`pip install $cmd\`"
 done
 
 PACKAGE="audioio"
 PACKAGEROOT="$(dirname "$(realpath "$0")")"
 BUILDROOT="$PACKAGEROOT/site"
+
+# check for code coverage report:
+# need to call nosetest with --with-coverage --cover-html --cover-xml
+HAS_COVER=false
+test -d cover && HAS_COVER=true
 
 echo
 echo "Clean up documentation of $PACKAGE"
@@ -18,22 +24,44 @@ echo
 rm -rf "$BUILDROOT" 2> /dev/null || true
 mkdir -p "$BUILDROOT"
 
-echo "Building general documentation for $PACKAGE"
-echo
+if command -v mkdocs >/dev/null; then
+    echo
+    echo "Building general documentation for $PACKAGE"
+    echo
 
-cd "$PACKAGEROOT"
-mkdocs build --config-file .mkdocs.yml --site-dir "$BUILDROOT" 
-cd - > /dev/null
+    cd "$PACKAGEROOT"
+    cp .mkdocs.yml mkdocs-tmp.yml
+    if $HAS_COVER; then
+	echo "    - Coverage: 'cover/index.html'" >> mkdocs-tmp.yml
+    fi
+    mkdocs build --config-file mkdocs-tmp.yml --site-dir "$BUILDROOT"
+    cd - > /dev/null
+fi
 
-echo
-echo "Building API reference docs for $PACKAGE"
-echo
+if $HAS_COVER; then
+    echo
+    echo "Copy code coverage report and generate badge for $PACKAGE"
+    echo
 
-cd "$PACKAGEROOT"
-pdoc3 --html --config sort_identifiers=False --output-dir "$BUILDROOT/api-tmp" $PACKAGE
-mv "$BUILDROOT/api-tmp/$PACKAGE" "$BUILDROOT/api"
-rmdir "$BUILDROOT/api-tmp"
-cd - > /dev/null
+    cd "$PACKAGEROOT"
+    cp -r cover "$BUILDROOT/"
+    genbadge coverage -i coverage.xml
+    # https://smarie.github.io/python-genbadge/
+    mv coverage-badge.svg site/coverage.svg
+    cd - > /dev/null
+fi
+
+if command -v pdoc3 >/dev/null; then
+    echo
+    echo "Building API reference docs for $PACKAGE"
+    echo
+
+    cd "$PACKAGEROOT"
+    pdoc3 --html --config latex_math=True --config sort_identifiers=False --output-dir "$BUILDROOT/api-tmp" $PACKAGE
+    mv "$BUILDROOT/api-tmp/$PACKAGE" "$BUILDROOT/api"
+    rmdir "$BUILDROOT/api-tmp"
+    cd - > /dev/null
+fi
 
 echo
 echo "Done. Docs in:"
